Started to migrate parcelset to HPX. Decoupled threadmanager from parcelset (nice!). Started to add Linux build system to HPX (but probably non-functional for now).
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@507 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ hpx/util/portable_binary_iarchive.hpp
@@ -143,32 +143,32 @@ public:
     void load(unsigned short & t){
         boost::intmax_t l;
         load_impl(l, sizeof(unsigned short));
-        t = l;
+        t = static_cast<unsigned short>(l);
     }
     void load(short & t){
         boost::intmax_t l;
         load_impl(l, sizeof(short));
-        t = l;
+        t = static_cast<short>(l);
     }
     void load(unsigned int & t){
         boost::intmax_t l;
         load_impl(l, sizeof(unsigned int));
-        t = l;
+        t = static_cast<unsigned int>(l);
     }
     void load(int & t){
         boost::intmax_t l;
         load_impl(l, sizeof(int));
-        t = l;
+        t = static_cast<int>(l);
     }
     void load(unsigned long & t){
         boost::intmax_t l;
         load_impl(l, sizeof(unsigned long));
-        t = l;
+        t = static_cast<unsigned long>(l);
     }
     void load(long & t){
         boost::intmax_t l;
         load_impl(l, sizeof(long));
-        t = l;
+        t = static_cast<long>(l);
     }
 #if defined(BOOST_HAS_LONG_LONG)
     void load(boost::long_long_type & t){
